+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 * Linters names no longer have a `Linter` suffix
 * Teach scss-lint that `@extend` should use placeholder selectors
 * Fix bug where a lint would be reported for hyphened keyword arguments
+* Teach scss-lint to recognize vendor-prefixed properties when enforcing sort
+  order
 
 0.7.1
 * Upgrade Sass gem dependency from 3.2.8 -> 3.2.9